fix: renaming several files for new convention
