+++ src/components/games/deal-or-no-deal/Board.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Paper from "@mui/material/Paper";
 import Case from "./Case";
-import { Briefcase } from "../../../recoil/deal-or-no-deal-atom";
+import { Briefcase } from "../../../recoil/deal-or-no-deal-state";
 
 interface BoardProps {
   board: Briefcase[];
