+++ src/components/games/deal-or-no-deal/Header.tsx
@@ -3,7 +3,7 @@ import Button from "@mui/material/Button";
 import Grid from "@mui/material/Grid";
 import Typography from "@mui/material/Typography";
 import { getMoneyText } from "./helpers";
-import { Briefcase } from "../../../recoil/deal-or-no-deal-atom";
+import { Briefcase } from "../../../recoil/deal-or-no-deal-state";
 import { DBPlayer } from "../../../recoil/player-atom";
 
 interface HeaderProps {
